requirements.txt
# YOLO + zależności do detekcji ramek
ultralytics==8.3.193
pillow==10.4.0

# PyTorch (CPU) kompatybilny z NumPy 1.x
torch==2.8.0
torchvision==0.23.0

# Wymuszenie NumPy <2, aby uniknąć błędów z modułami skompilowanymi pod 1.x
numpy<2

# OpenAI API
openai>=1.0.0